Create CelebrityProblem.java
A celebrity is a person who is known to all but does not know anyone at a party.A square NxN matrix M[][] is used to represent people at the party such that if an element of row i and column j  is set to 1 it means ith person knows jth person. Here M[i][i] will always be 0.